requirements.txt
python-dotenv>=1.0.1,<1.1
httpx>=0.26.0,<0.27
